requirements.txt
#test
Flask==3.0.3
Flask-Login==0.6.3
fpdf2==2.7.9
SQLAlchemy==2.0.35
DateTime==5.5
pandas==2.2.3
matplotlib==3.9.2
